+++ 3-Greedy Algorithms, Minimum Spanning Tree, and Dynamic Programming/2-Dynamic Programming/2-Knapsack Problem/knapsack.py
@@ -35,7 +35,7 @@ def knapsack(vals, weights, cap):
     """
     Solves the knapsack problem of the items with the given values and weights,
     and the given capacity, in an improved bottom-up way.
-    :param vals: list[int]
+    :param vals: list[float]
     :param weights: list[int]
     :param cap: int
     :return: set{int}
@@ -73,7 +73,7 @@ def _reconstruct(vals, weights, cap, subproblems):
     """
     Private helper function to reconstruct the included items according to the
     optimal solution using backtracking.
-    :param vals: list[int]
+    :param vals: list[float]
     :param weights: list[int]
     :param cap: int
     :param subproblems: list[list[int]]
